AT-152: move jobs to module
