+++ frontend/src/app/core/main/main.module.ts
@@ -3,9 +3,10 @@ import { CommonModule } from '@angular/common';
 import { ToolbarModule } from './toolbar/toolbar.module';
 import { MainComponent } from './main.component';
 import { RouterOutlet } from '@angular/router';
+import { MainRoutingModule } from './main-routing.module';
 
 @NgModule({
   declarations: [MainComponent],
-  imports: [CommonModule, ToolbarModule, RouterOutlet]
+  imports: [CommonModule, MainRoutingModule, ToolbarModule, RouterOutlet]
 })
 export class MainModule {}
