+++ frontend/src/app/feature/jobs/jobs-routing.module.ts
@@ -2,14 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { JobsComponent } from './jobs.component';
 
-const routes: Routes = [
-  { path: '', component: JobsComponent },
-  {
-    path: 'jobs',
-    loadChildren: () =>
-      import('./search/search.module').then((m) => m.SearchModule)
-  }
-];
+const routes: Routes = [{ path: '', component: JobsComponent }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
